requirements.txt
gradio>=4.0.0
websocket-client>=1.6.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.28.0
cryptography>=41.0.0

# Optional: Advanced audio analysis for Lipsync segmentation (beat detection)
# Install with: pip install librosa
librosa>=0.10.0
